requirements.txt
# Pair Trading System - Complete Requirements

# ============================================================
# Core Trading & Market Data
# ============================================================
MetaTrader5>=5.0.45
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ============================================================
# Statistics & Machine Learning
# ============================================================
statsmodels>=0.14.0
scikit-learn>=1.3.0
arch>=6.2.0

# ============================================================
# GUI Framework (PyQt6)
# ============================================================
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0

# ============================================================
# Configuration Management
# ============================================================
PyYAML>=6.0
python-dotenv>=1.0.0

# ============================================================
# Charting & Visualization (Optional)
# ============================================================
matplotlib>=3.7.0
pyqtgraph>=0.13.0
plotly>=5.17.0

# ============================================================
# Utilities
# ============================================================
python-dateutil>=2.8.0
pytz>=2023.3
colorama>=0.4.6

# ============================================================
# Testing & Development
# ============================================================
pytest>=7.4.0
pytest-cov>=4.1.0
